+++ boa/contracts/vyper/decoder_utils.py
@@ -14,6 +14,33 @@ from vyper.semantics.types import (
 )
 from vyper.utils import unsigned_to_signed
 
+from boa.util.abi import Address
+from boa.vm.utils import ceil32, floor32
+
+
+class ByteAddressableStorage:
+    def __init__(self, evm, address: Address, key: int):
+        self.evm = evm
+        self.address = address.canonical_address
+        self.key = key
+
+    def __getitem__(self, subscript):
+        if isinstance(subscript, slice):
+            ret = b""
+            start = subscript.start or 0
+            stop = subscript.stop
+            i = self.key + start // 32
+            while i < self.key + ceil32(stop) // 32:
+                ret += self.evm.get_storage_slot(self.address, i)
+                i += 1
+
+            start_ofst = floor32(start)
+            start -= start_ofst
+            stop -= start_ofst
+            return memoryview(ret[start:stop])
+        else:
+            raise Exception("Must slice {self}")
+
 
 class _Struct(dict):
     def __init__(self, name, *args, **kwargs):
